requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0